img_webapp/requirements.txt
mysql-connector-python
flask
redis
